Remove markdown renderer from ChatMessage
Removed the chat message markdown renderer as part of the refactoring process. As we want to keep our chatbot responses simple, it was found that markdown rendering was not necessary. The ChatbotSettings was imported in various files because it's needed for new feature implementations. Configurations in rollup.config.mjs and tsconfig were made to resolve some bundling and transpiling issues. Also, removed react-markdown from package-lock.json and replaced it with markdown-to-jsx as it better suits our needs.