Updated "id" type in multiple instances
The type of "id" value was changed from number to string in database.types.ts and mutations.ts files. This was required to allow for more flexibility when assigning IDs in different scenarios. Also, an irrelevant line was removed to streamline the content of the CrawlWebsiteModal.tsx file. The text was updated to improve the user guide for disallowing URLs.