Modify codebase to support I18n (translations)
The changes were made across multiple files in the codebase to allow support for I18n. This essentially allows for text to be internationalized which means it can be translated into multiple languages based on locale. This was done to make the application accessible to a wider audience by catering for users that speak languages other than English. Changes included adding translation methods for static strings in the code and updating the return types of some properties to facilitate dynamic values for different languages.