Add data-loader-supabase-nextjs and refactor database loading
Integrated @makerkit/data-loader-supabase-nextjs to simplify and optimize the loading of data from the database. The previous manual loading method getJobs() was removed, and ServerDataLoader was implemented to perform the same task in a more effective way. The fetch behavior in the JobsTable and ChatbotLayout components was also adjusted to fit the new data-loader usage. Updated dependency versions in package.json and package-lock.json suitable with this change.